update notebook 7
