+++ flodym/flow_helper.py
@@ -1,7 +1,7 @@
 from typing import Callable
 
 from .processes import Process
-from .named_dim_arrays import Flow
+from .flodym_arrays import Flow
 from .dimensions import DimensionSet
 from .mfa_definition import FlowDefinition
 from .flow_naming import process_names_with_arrow
