+++ howtos/input_data/array_data.csv
@@ -1,0 +1,7 @@
+Time,Material,value
+2010,Fe,1.0
+2020,Fe,1.0
+2010,Cu,1.0
+2020,Cu,1.0
+2010,Mn,1.0
+2020,Mn,1.0
